6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
+
 /**
 * main - Prints the size of various  types a based on
 * the computer it is compiled and run on. .
-* Return: Always 0.
+* Return 0 (Success)
  */
 int main(void)
 {
-printf("Size of char: %zu byte(s)\n", sizeof(char));
-printf("Size of int: %zu bytes(s)\n", sizeof(int));
-printf("Size of long int: %zu bytes(s)\n", sizeof(long int));
-printf("Size of long long int: %zu  bytes(s)\n", sizeof(long long int));
- printf("Size of float: %zu bytes(s)\n", sizeof(float));   
+printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(char));
+printf("Size of int: %lu bytes(s)\n", (unsigned long)sizeof(int));
+printf("Size of long int: %lu bytes(s)\n", (unsigned long)sizeof(long int));
+printf("Size of long long int: %lu  bytes(s)\n", (unsigned long)sizeof(long long int));
+ printf("Size of float: %lu bytes(s)\n", (unsigned long)sizeof(float));   
 return (0);
 }
-
